Added normal RBM

+++ src/main/java/InOutOperations.java
@@ -1,9 +1,3 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
 import java.io.File;
 import java.io.IOException;
 import java.io.ObjectInputStream;
@@ -13,15 +7,20 @@ import java.text.SimpleDateFormat;
 import java.util.Date;
 import org.apache.commons.io.FileUtils;
 
+/**
+ * DataConverter
+ *
+ * @author Radek
+ */
 public class InOutOperations {
     private static final String simpleWeightsFolder = "Output/SimpleWeights";
     private static final String imageExportFolder = "Output/ImageExport"; 
     
-    public static void saveSimpleWeights(float[][] weights, Date date) throws IOException{
+    public static void saveSimpleWeights(double[][] weights, Date date) throws IOException{
         saveSimpleWeights(weights, date, "weights");
     }
     
-    public static void saveSimpleWeights(float[][] weights, Date date, String suffix) throws IOException{
+    public static void saveSimpleWeights(double[][] weights, Date date, String suffix) throws IOException{
         mkdir(simpleWeightsFolder);
         File file = new File(simpleWeightsFolder + "/" + getFileNameByDate(date, suffix, "dat"));
         ObjectOutputStream oos = new ObjectOutputStream(Files.newOutputStream(file.toPath()));
@@ -29,9 +28,9 @@ public class InOutOperations {
         oos.close();
     }
     
-    public static float[][] loadSimpleWeights(File file) throws IOException, ClassNotFoundException{
+    public static double[][] loadSimpleWeights(File file) throws IOException, ClassNotFoundException{
 	ObjectInputStream ois = new ObjectInputStream(Files.newInputStream(file.toPath()));
-        float[][] weights = (float[][]) ois.readObject();
+        double[][] weights = (double[][]) ois.readObject();
         ois.close();
         return weights; 
     }
